Extract ConditionsInputs shared component
